requirements.txt
appdirs==1.4.3
attrs==19.3.0
Babel==2.8.0
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
defusedxml==0.6.0
Deprecated==1.2.10
distlib==0.3.0
distro==1.4.0
Django==2.2.14
django-allauth==0.42.0
django-authtools==1.7.0
django-braces==1.14.0
django-ckeditor==5.9.0
django-environ==0.4.5
django-filter==2.2.0
django-jet==1.0.8
django-js-asset==1.2.2
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
django-solo==1.1.3
djangorestframework==3.11.0
drf-yasg==1.17.1
Faker==0.9.1
fcm-django==0.3.4
google-api-core==1.22.1
google-api-python-client==1.11.0
google-auth==1.21.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.1
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.18.1
idna==2.8
importlib-metadata==1.7.0
inflection==0.5.0
iniconfig==1.0.1
install==1.3.3
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
mixer==6.1.3
more-itertools==8.4.0
msgpack==0.6.2
mysqlclient==2.0.1
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
phonenumberslite==8.12.7
Pillow==7.2.0
pluggy==0.13.1
progress==1.5
protobuf==3.13.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.7
PyJWT==1.7.1
pyparsing==2.4.6
pytest==6.0.1
pytest-django==3.9.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
qrcode==6.1
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sentry-sdk==0.17.1
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.2
toml==0.10.1
twilio==6.45.0
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.12.1
zipp==3.1.0
